refactor: move PartialModel functions in separate object, test: add PartialModelUtilsTest
